Changes to PersistCity & PersistWeatherData
PeristCity is now searching for a entry in the DB with the same Zip Code
first before persisting. If we have already persisted this city, the
city is skipped to prevent an error (Cities need to be unique in DB).

PersistWeatherData is now fetching the correct city from the DB based on
the City object that is defined within the WeatherData object. (Again
using em.find() with Zip Code. This makes sure that the relation between
the WeatherData entry and the correct City entry in the DB is properly
implemented.