Multiple smaller changes
BusinessHandlerImpl:
Changed selectWeatherByDateAndCity() method to automatically create the
CSV file when called. Additionally, only the first 10 Weatherdata
Objects from the List are being sent over to the UI. (Use them as a
preview for the User=

demoUI: Demoimplementation of the Businesshandler can be tested here.
Currently selecting for Zürich from 01.10.2021 until 31.12.2021 and only
showing the first 10 records. Weatherdata.csv should be created in the
Downloads folder with all records available.

WeatherData: Smaller formatting change to toString()

DbHelper: Sorted the result of selectWeatherDataSingleCity() by timestamp ascending so that the data is ordered in the CSV-File