Finalized CRUD implementation together with Nilukzil
Implementation of delete and Update methods has been done.
User is getting an incremental number at the end of their username if
there is already an existing user in the DB with the same Userid.

All necessary checks are performed when calling updateUser(), no further
checks needed. Always make sure to call updateUser() with the new User
object without updating the Userid at first since the old Userid is
needed to fetch the already existing Entity in the DB.