Added TestCases for User statements
Added TestCases for update and delete.
Changed the way UserId's get created for Lastnames > 5 letters.
Then the lastname gets fully taken instead of only 5 letters from it.
The checkExistingUser is now called in the insertUser method instead of
the update method. To check if the User exists(and recreate the ID)
already in the insert method.